# artifacts/api-server/src/routes/noble_score.ts
import { Router, type Request } from "express";
import { db } from "@workspace/db";
import { usersTable, zuil_voortgangTable, nobleScoreLogTable, scenariosTable, userUseCaseRatingsTable, DEFAULT_BEHAVIOR_PROFILE, type BehaviorProfile, type ScenarioContent, type WardrobeItem, getPillarName } from "@workspace/db";
import { eq, and, desc, inArray } from "drizzle-orm";
import { z } from "zod";
import { extractToken } from "../lib/auth-middleware";

const router = Router();

const LogQuerySchema = z.object({
  limit: z.coerce.number().int().min(1).max(100).default(20),
});

const SocialClassSchema = z.enum(["universal", "elite", "middle_class"]).default("universal");

/**
 * Build the pillar name/domain array for a given social class register.
 * Falls back to "universal" when the requested class has no overrides.
 */
function getPillarNamesForClass(socialClass: "universal" | "elite" | "middle_class") {
  return [1, 2, 3, 4, 5].map((p) => ({
    pillar: p,
    ...getPillarName(p, socialClass),
  }));
}

/**
 * Resolves the authenticated user ID from the session cookie (or Bearer fallback).
 * Returns null if no token is present or if the token is not recognised.
 * Used by endpoints that should work for both guests and authenticated users.
 */
async function optionalUserFromToken(req: Request): Promise<string | null> {
  const token = extractToken(req);
  if (!token) return null;
  const [user] = await db
    .select({ id: usersTable.id })
    .from(usersTable)
    .where(eq(usersTable.session_token, token))
    .limit(1);
  return user?.id ?? null;
}

/** Default (universal) pillar name list — used for initialising pillar progress rows. */
const PILLAR_NAMES = getPillarNamesForClass("universal");

const PILLAR_TITLES: Record<number, string[]> = {
  1: ["The Observer", "The Reader", "The Traveller", "The Diplomat", "The Luminary"],
  2: ["The Conscious", "The Poised", "The Curated", "The Impeccable", "The Icon"],
  3: ["The Listener", "The Conversant", "The Eloquent", "The Orator", "The Sage"],
  4: ["The Guest", "The Diner", "The Connoisseur", "The Host", "The Maître"],
  5: ["The Sipper", "The Taster", "The Selector", "The Sommelier", "The Cellar Master"],
};

const GLOBAL_LEVELS = [
  { min: 0, max: 19, level: 1, name: "The Aware", color: "#9E9E9E" },
  { min: 20, max: 39, level: 2, name: "The Composed", color: "#26A69A" },
  { min: 40, max: 59, level: 3, name: "The Refined", color: "#7E57C2" },
  { min: 60, max: 79, level: 4, name: "The Distinguished", color: "#FFA726" },
  { min: 80, max: 100, level: 5, name: "The Sovereign", color: "#AD1457" },
];

function getLevelFromScore(score: number) {
  const normalised = Math.min(100, Math.max(0, score));
  return GLOBAL_LEVELS.find(l => normalised >= l.min && normalised <= l.max) || GLOBAL_LEVELS[0];
}

function getTitleFromScore(pillar: number, score: number): string {
  const titles = PILLAR_TITLES[pillar] || [];
  const level = getLevelFromScore(score).level - 1;
  return titles[level] || titles[0] || "The Aware";
}

function getNextTitle(pillar: number, score: number): string | null {
  const titles = PILLAR_TITLES[pillar] || [];
  const level = getLevelFromScore(score).level;
  return titles[level] || null;
}

function getProgressPercent(score: number): number {
  const level = getLevelFromScore(score);
  const range = level.max - level.min;
  const progress = score - level.min;
  return Math.round((progress / range) * 100);
}

async function ensurePillarProgress(userId: string) {
  const existing = await db.select().from(zuil_voortgangTable).where(eq(zuil_voortgangTable.user_id, userId));
  const existingPillars = new Set(existing.map(e => e.pillar));

  const toCreate = PILLAR_NAMES.filter(p => !existingPillars.has(p.pillar));

  if (toCreate.length > 0) {
    await db.insert(zuil_voortgangTable).values(
      toCreate.map(p => ({
        user_id: userId,
        pillar: p.pillar,
        score: 0,
        current_title: PILLAR_TITLES[p.pillar][0],
      }))
    );
  }
}

router.get("/noble-score", async (req, res) => {
  try {
    const userId = await optionalUserFromToken(req);
    const socialClass = SocialClassSchema.catch("universal").parse(req.query.social_class);
    const pillarNames = getPillarNamesForClass(socialClass);

    if (!userId) {
      const emptyPillars = pillarNames.map(p => ({
        pillar: p.pillar,
        pillar_name: p.pillar_name,
        pillar_domain: p.pillar_domain,
        score: 0,
        current_title: getTitleFromScore(p.pillar, 0),
        next_title: getNextTitle(p.pillar, 0),
        progress_percent: 0,
      }));
      return res.json({
        total_score: 0,
        level: 1,
        level_name: GLOBAL_LEVELS[0].name,
        level_color: GLOBAL_LEVELS[0].color,
        next_level_threshold: GLOBAL_LEVELS[1].min,
        next_level_name: GLOBAL_LEVELS[1].name,
        pillars: emptyPillars,
      });
    }

    await ensurePillarProgress(userId);

    const [user] = await db.select().from(usersTable).where(eq(usersTable.id, userId)).limit(1);
    const pillarRows = await db.select().from(zuil_voortgangTable).where(eq(zuil_voortgangTable.user_id, userId));

    const totalScore = user ? user.noble_score : 0;
    const currentLevel = getLevelFromScore(totalScore);
    const nextLevel = GLOBAL_LEVELS.find(l => l.level === currentLevel.level + 1);

    const pillars = pillarNames.map(p => {
      const row = pillarRows.find(r => r.pillar === p.pillar);
      const score = row?.score ?? 0;
      return {
        pillar: p.pillar,
        pillar_name: p.pillar_name,
        pillar_domain: p.pillar_domain,
        score,
        current_title: getTitleFromScore(p.pillar, score),
        next_title: getNextTitle(p.pillar, score),
        progress_percent: getProgressPercent(score),
      };
    });

    return res.json({
      total_score: totalScore,
      level: currentLevel.level,
      level_name: currentLevel.name,
      level_color: currentLevel.color,
      next_level_threshold: nextLevel?.min ?? 100,
      next_level_name: nextLevel?.name ?? null,
      pillars,
    });
  } catch (err) {
    req.log.error({ err }, "Failed to fetch noble score");
    return res.status(500).json({ error: "Your standing could not be retrieved at this moment." });
  }
});

const SubmitAnswerBodySchema = z.object({
  scenario_id: z.number().int().positive(),
  selected_option_index: z.number().int().min(0),
  time_taken_seconds: z.number().int().optional(),
  lang: z.string().min(2).max(10).optional(),
});

type AgeBand = "young" | "adult" | "mature";

function getAgeBand(birthYear: number | null | undefined): AgeBand {
  if (!birthYear) return "adult";
  const age = new Date().getFullYear() - birthYear;
  if (age <= 28) return "young";
  if (age <= 50) return "adult";
  return "mature";
}

/** i18n key pools — the frontend translates these via t(key). */
const MENTOR_FEEDBACK_CORRECT: Record<AgeBand, string[]> = {
  young:  ["mentor.correct.young.0", "mentor.correct.young.1", "mentor.correct.young.2", "mentor.correct.young.3"],
  adult:  ["mentor.correct.adult.0", "mentor.correct.adult.1", "mentor.correct.adult.2", "mentor.correct.adult.3"],
  mature: ["mentor.correct.mature.0", "mentor.correct.mature.1", "mentor.correct.mature.2", "mentor.correct.mature.3"],
};

const MENTOR_FEEDBACK_INCORRECT: Record<AgeBand, string[]> = {
  young:  ["mentor.incorrect.young.0", "mentor.incorrect.young.1", "mentor.incorrect.young.2", "mentor.incorrect.young.3"],
  adult:  ["mentor.incorrect.adult.0", "mentor.incorrect.adult.1", "mentor.incorrect.adult.2", "mentor.incorrect.adult.3"],
  mature: ["mentor.incorrect.mature.0", "mentor.incorrect.mature.1", "mentor.incorrect.mature.2", "mentor.incorrect.mature.3"],
};

function clamp(n: number, min = 0, max = 100): number {
  return Math.min(max, Math.max(min, n));
}

function updateBehaviorProfile(
  current: BehaviorProfile,
  boltonCluster: number | null | undefined,
  isCorrect: boolean,
  behaviorSignal: string | null | undefined,
): BehaviorProfile {
  const profile: BehaviorProfile = JSON.parse(JSON.stringify(current));
  const delta = isCorrect ? 4 : -3;

  if (boltonCluster === 1) {
    profile.listening_score = clamp(profile.listening_score + delta);
    profile.eq_dimensions.self_awareness = clamp(profile.eq_dimensions.self_awareness + Math.round(delta * 0.5));
    profile.eq_dimensions.social_skill = clamp(profile.eq_dimensions.social_skill + Math.round(delta * 0.5));
  } else if (boltonCluster === 2) {
    profile.eq_dimensions.self_regulation = clamp(profile.eq_dimensions.self_regulation + delta);
    const assertivenessSignals: BehaviorProfile["assertiveness_style"][] = ["assertive", "passive", "aggressive", "passive_aggressive"];
    const resolvedStyle = behaviorSignal && assertivenessSignals.includes(behaviorSignal as BehaviorProfile["assertiveness_style"])
      ? (behaviorSignal as BehaviorProfile["assertiveness_style"])
      : isCorrect ? "assertive" : "passive";
    profile.assertiveness_style = resolvedStyle;
  } else if (boltonCluster === 3) {
    profile.eq_dimensions.empathy = clamp(profile.eq_dimensions.empathy + delta);
    const conflictSignals: BehaviorProfile["conflict_mode"][] = ["collaborate", "avoid", "compete", "accommodate"];
    const resolvedMode = behaviorSignal && conflictSignals.includes(behaviorSignal as BehaviorProfile["conflict_mode"])
      ? (behaviorSignal as BehaviorProfile["conflict_mode"])
      : isCorrect ? "collaborate" : "avoid";
    profile.conflict_mode = resolvedMode;
  } else {
    profile.eq_dimensions.social_skill = clamp(profile.eq_dimensions.social_skill + Math.round(delta * 0.5));
    profile.nonverbal_awareness = clamp(profile.nonverbal_awareness + Math.round(delta * 0.3));
  }

  return profile;
}

router.post("/noble-score/submit", async (req, res) => {
  try {
    const userId = await optionalUserFromToken(req);

    const bodyParsed = SubmitAnswerBodySchema.safeParse(req.body);
    if (!bodyParsed.success) {
      return res.status(400).json({ error: "The submission does not meet the expected form. Please review and resubmit." });
    }

    const { scenario_id, selected_option_index, time_taken_seconds, lang } = bodyParsed.data;

    const [scenario] = await db.select().from(scenariosTable).where(eq(scenariosTable.id, scenario_id)).limit(1);

    if (!scenario) {
      return res.status(404).json({ error: "The referenced scenario could not be located." });
    }

    const content = scenario.content_json;
    const selectedOption = content.options[selected_option_index];

    if (!selectedOption) {
      return res.status(400).json({ error: "The selected option is not valid for this scenario." });
    }

    /**
     * Correctness resolution — backwards-compatible.
     * New content uses answer_tier (1 = Correct, 2 = Acceptable, 3 = Incorrect).
     * Legacy content uses correct: boolean.
     * isCorrect = true for tier 1 only; tier 2 is partially rewarded below.
     */
    const resolvedTier: 1 | 2 | 3 = selectedOption.answer_tier
      ?? (selectedOption.correct === true ? 1 : 3);
    const isCorrect = resolvedTier === 1;
    const isAcceptable = resolvedTier === 2;

    let scoreDelta: number;
    if (isCorrect) {
      scoreDelta = scenario.noble_score_impact;
    } else if (isAcceptable) {
      scoreDelta = Math.floor(scenario.noble_score_impact * 0.5);
    } else {
      scoreDelta = Math.floor(-scenario.noble_score_impact * 0.2);
    }

    if (isCorrect && time_taken_seconds && time_taken_seconds < 10) {
      scoreDelta = Math.round(scoreDelta * 1.2);
    }

    const trigger = (isCorrect || isAcceptable) ? "correct_choice" : "incorrect_choice";

    let newTotalScore = 0;
    let levelUp = false;
    let newLevelName: string | null = null;
    let ageBand: AgeBand = "adult";

    if (userId) {
      await ensurePillarProgress(userId);

      const [pillarRow] = await db.select().from(zuil_voortgangTable)
        .where(and(
          eq(zuil_voortgangTable.user_id, userId),
          eq(zuil_voortgangTable.pillar, scenario.pillar),
        ))
        .limit(1);

      const oldPillarScore = pillarRow?.score ?? 0;
      const newPillarScore = Math.min(100, Math.max(0, oldPillarScore + scoreDelta));
      const newTitle = getTitleFromScore(scenario.pillar, newPillarScore);

      await db.update(zuil_voortgangTable)
        .set({ score: newPillarScore, current_title: newTitle })
        .where(and(
          eq(zuil_voortgangTable.user_id, userId),
          eq(zuil_voortgangTable.pillar, scenario.pillar),
        ));

      const [user] = await db.select().from(usersTable).where(eq(usersTable.id, userId)).limit(1);
      const oldTotalScore = user?.noble_score ?? 0;
      newTotalScore = Math.min(100, Math.max(0, oldTotalScore + Math.round(scoreDelta * 0.5)));

      const oldLevel = getLevelFromScore(oldTotalScore);
      const newLevel = getLevelFromScore(newTotalScore);
      levelUp = newLevel.level > oldLevel.level;
      newLevelName = levelUp ? newLevel.name : null;

      if (user) {
        const profilingEnabled = (user as { profiling_consent?: boolean | null }).profiling_consent !== false;
        const currentProfile: BehaviorProfile = (user as { behavior_profile?: BehaviorProfile | null }).behavior_profile ?? DEFAULT_BEHAVIOR_PROFILE;
        const scenarioContent = scenario.content_json as ScenarioContent | null;
        const behaviorSignal = scenarioContent?.options?.[selected_option_index]?.behavior_signal ?? null;
        const updatedProfile = profilingEnabled
          ? updateBehaviorProfile(
              currentProfile,
              (scenario as { bolton_cluster?: number | null }).bolton_cluster,
              isCorrect,
              behaviorSignal,
            )
          : currentProfile;

        // Streak logic
        const today = new Date().toISOString().split("T")[0];
        const lastActivity = (user as { last_activity_date?: string | null }).last_activity_date;
        const yesterday = new Date(Date.now() - 86400000).toISOString().split("T")[0];
        const currentStreak = (user as { daily_streak?: number }).daily_streak ?? 0;
        let newStreak = currentStreak;
        if (lastActivity === today) {
          newStreak = currentStreak;
        } else if (lastActivity === yesterday) {
          newStreak = currentStreak + 1;
        } else {
          newStreak = 1;
        }

        // Wardrobe unlock logic — Pillar 2 (Appearance)
        const currentWardrobe: WardrobeItem[] = (user as { wardrobe_unlocks?: WardrobeItem[] | null }).wardrobe_unlocks ?? [];
        let newWardrobe = currentWardrobe;
        if (scenario.pillar === 2 && isCorrect) {
          const WARDROBE_THRESHOLDS = [
            { minScore: 20, id: "italian_suit", name: "Italian Suit", region: "IT" },
            { minScore: 40, id: "arabic_thobe", name: "Arabic Thobe", region: "SA" },
            { minScore: 60, id: "japanese_hakama", name: "Japanese Hakama", region: "JP" },
            { minScore: 80, id: "scottish_tartan", name: "Scottish Tartan", region: "GB" },
          ];
          const alreadyUnlocked = new Set(currentWardrobe.map((w) => w.id));
          for (const threshold of WARDROBE_THRESHOLDS) {
            if (newPillarScore >= threshold.minScore && !alreadyUnlocked.has(threshold.id)) {
              newWardrobe = [...newWardrobe, {
                id: threshold.id,
                name: threshold.name,
                region: threshold.region,
                pillar: 2,
                unlocked_at: today,
              }];
              alreadyUnlocked.add(threshold.id);
            }
          }
        }

        // Avatar state — update based on noble score level
        const newAvatarState = {
          rank_badge: getLevelFromScore(newTotalScore).name,
          style_tier: getLevelFromScore(newTotalScore).level,
        };

        await db.update(usersTable)
          .set({
            noble_score: newTotalScore,
            ...(profilingEnabled && { behavior_profile: updatedProfile }),
            daily_streak: newStreak,
            last_activity_date: today,
            wardrobe_unlocks: newWardrobe,
            avatar_state: newAvatarState,
          })
          .where(eq(usersTable.id, userId));
      }

      await db.insert(nobleScoreLogTable).values({
        user_id: userId,
        scenario_id,
        score_delta: scoreDelta,
        trigger,
        level_name_after: newLevelName,
      });

      // Invalidate use-case readiness cache so the next /use-cases request
      // recomputes fresh scores that reflect the updated pillar progress.
      // Isolated in its own try/catch so a transient cache-delete failure
      // does not turn a successful scenario submission into a 500 response.
      try {
        await db.delete(userUseCaseRatingsTable).where(
          eq(userUseCaseRatingsTable.user_id, userId)
        );
      } catch (cacheErr) {
        req.log.warn({ cacheErr }, "Could not invalidate use-case readiness cache after scenario submit");
      }

      ageBand = getAgeBand(user?.birth_year);
    }

    const correctionStyle = (scenario as { correction_style?: string | null }).correction_style;
    const requestLang = lang ? lang.split("-")[0] : "en";
    const isEnglish = !lang || requestLang === "en";

    let mentorFeedback: string;
    if (!isCorrect && correctionStyle && isEnglish) {
      // Bespoke English correction tip — only shown to English-locale users.
      // Non-English users receive a translated i18n key from the pool below.
      mentorFeedback = correctionStyle;
    } else {
      const feedbackPool = isCorrect
        ? MENTOR_FEEDBACK_CORRECT[ageBand]
        : MENTOR_FEEDBACK_INCORRECT[ageBand];
      mentorFeedback = feedbackPool[Math.floor(Math.random() * feedbackPool.length)];
    }

    return res.json({
      correct: isCorrect,
      answer_tier: resolvedTier,
      explanation: selectedOption.explanation,
      mentor_feedback: mentorFeedback,
      score_delta: scoreDelta,
      new_total_score: newTotalScore,
      level_up: levelUp,
      new_level_name: newLevelName,
    });
  } catch (err) {
    req.log.error({ err }, "Failed to submit scenario answer");
    return res.status(500).json({ error: "A difficulty arose while processing your response. Please try again." });
  }
});

router.get("/noble-score/log", async (req, res) => {
  try {
    const userId = await optionalUserFromToken(req);
    if (!userId) return res.json([]);

    const parsed = LogQuerySchema.safeParse(req.query);
    const limit = parsed.success ? parsed.data.limit : 20;

    const log = await db.select()
      .from(nobleScoreLogTable)
      .where(eq(nobleScoreLogTable.user_id, userId))
      .orderBy(desc(nobleScoreLogTable.timestamp))
      .limit(limit);

    const scenarioIds = [...new Set(log.map(l => l.scenario_id).filter((id): id is number => id !== null))];

    let scenarioMap: Record<number, { title: string; pillar: number; pillar_domain: string }> = {};
    if (scenarioIds.length > 0) {
      const scenarios = await db.select({
        id: scenariosTable.id,
        title: scenariosTable.title,
        pillar: scenariosTable.pillar,
        social_class: scenariosTable.social_class,
      }).from(scenariosTable).where(inArray(scenariosTable.id, scenarioIds));

      for (const s of scenarios) {
        const sc = (s.social_class as "universal" | "elite" | "middle_class") ?? "universal";
        const { pillar_domain } = getPillarName(s.pillar, sc);
        scenarioMap[s.id] = {
          title: s.title,
          pillar: s.pillar,
          pillar_domain,
        };
      }
    }

    const enriched = log.map(entry => ({
      ...entry,
      scenario_title: entry.scenario_id != null ? (scenarioMap[entry.scenario_id]?.title ?? null) : null,
      scenario_pillar: entry.scenario_id != null ? (scenarioMap[entry.scenario_id]?.pillar ?? null) : null,
      scenario_pillar_domain: entry.scenario_id != null ? (scenarioMap[entry.scenario_id]?.pillar_domain ?? null) : null,
    }));

    return res.json(enriched);
  } catch (err) {
    req.log.error({ err }, "Failed to fetch noble score log");
    return res.status(500).json({ error: "Your progress history is momentarily unavailable." });
  }
});

router.get("/noble-score/pillars", async (req, res) => {
  try {
    const userId = await optionalUserFromToken(req);
    const socialClass = SocialClassSchema.catch("universal").parse(req.query.social_class);
    const pillarNames = getPillarNamesForClass(socialClass);

    if (!userId) {
      return res.json(pillarNames.map(p => ({
        pillar: p.pillar,
        pillar_name: p.pillar_name,
        pillar_domain: p.pillar_domain,
        score: 0,
        current_title: getTitleFromScore(p.pillar, 0),
        next_title: getNextTitle(p.pillar, 0),
        progress_percent: 0,
      })));
    }

    await ensurePillarProgress(userId);

    const pillarRows = await db.select().from(zuil_voortgangTable).where(eq(zuil_voortgangTable.user_id, userId));

    const pillars = pillarNames.map(p => {
      const row = pillarRows.find(r => r.pillar === p.pillar);
      const score = row?.score ?? 0;
      return {
        pillar: p.pillar,
        pillar_name: p.pillar_name,
        pillar_domain: p.pillar_domain,
        score,
        current_title: getTitleFromScore(p.pillar, score),
        next_title: getNextTitle(p.pillar, score),
        progress_percent: getProgressPercent(score),
      };
    });

    return res.json(pillars);
  } catch (err) {
    req.log.error({ err }, "Failed to fetch pillar progress");
    return res.status(500).json({ error: "Your pillar progress is momentarily unavailable." });
  }
});

export default router;