# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOWISO Etiquette API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /noble-score/submit:
    post:
      operationId: submitScenarioAnswer
      tags: [noble-score]
      summary: Submit a scenario answer and update Noble Score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: Answer result with score update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /noble-score/log:
    get:
      operationId: getNobleScoreLog
      tags: [noble-score]
      summary: Get recent Noble Score events for the current user
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Score log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreLogEntry"
...
components:
  schemas:
...
    SubmitAnswerBody:
      type: object
      properties:
        scenario_id:
          type: integer
        selected_option_index:
          type: integer
        time_taken_seconds:
          type: integer
        lang:
          type: string
          description: BCP-47 language tag (e.g. "nl", "fr"). Used to return mentor feedback in the correct language.
      required:
        - scenario_id
        - selected_option_index
    AnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        explanation:
          type: string
        mentor_feedback:
          type: string
        score_delta:
          type: integer
        new_total_score:
          type: integer
        level_up:
          type: boolean
        new_level_name:
          type: string
          nullable: true
      required:
        - correct
        - explanation
        - mentor_feedback
        - score_delta
        - new_total_score
        - level_up
    ScoreLogEntry:
      type: object
      properties:
        id:
          type: integer
        scenario_id:
          type: integer
          nullable: true
        score_delta:
          type: integer
        trigger:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - score_delta
        - trigger
        - timestamp
...